pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cidoc-rag-assistant"
version = "0.1.0"
description = "FAISS-based RAG assistant for CIDOC CRM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "faiss-cpu>=1.8.0",
  "numpy>=1.26.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "rdflib>=7.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
